libft finito! ok supreme test

+++ School_42/libft/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: asantoro <asantoro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/09 12:54:28 by asantoro          #+#    #+#             */
-/*   Updated: 2022/03/10 12:00:58 by asantoro         ###   ########.fr       */
+/*   Updated: 2022/04/05 15:28:22 by asantoro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *str, unsigned int n)
+void	ft_bzero(void *str, size_t n)
 {
 	ft_memset(str, 0, n);
 }
